backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pyyaml==6.0.1
volcengine-python-sdk[ark]